+++ src/Tests/AllTests.fs
@@ -45,7 +45,9 @@ let allTests =
                 LogicApps.tests
                 Maps.tests
                 NetworkSecurityGroup.tests
+                OperationsManagement.tests
                 PostgreSQL.tests
+                PrivateLink.tests
                 ResourceGroup.tests
                 ServiceBus.tests
                 SignalR.tests